Make viewport width & height editable fields
