+++ packages/react-cosmos-playground2/src/shared/ui/valueInputTree/ValueInputTreeItem/NumberItem.tsx
@@ -1,13 +1,6 @@
 import React from 'react';
-import { KEY_DOWN, KEY_UP } from '../../../keys';
-import {
-  Label,
-  TextContainer,
-  TextField,
-  TextInputContainer,
-  TextMirror,
-  ValueContainer
-} from './shared';
+import { NumberInput } from '../../inputs/NumberInput';
+import { Label, ValueContainer } from './shared';
 
 type Props = {
   id: string;
@@ -17,59 +10,13 @@ type Props = {
 };
 
 export function NumberItem({ id, label, value, onChange }: Props) {
-  const [focused, setFocused] = React.useState(false);
-  const onFocus = React.useCallback(() => setFocused(true), []);
-  const onBlur = React.useCallback(() => setFocused(false), []);
-
-  const onInputChange = React.useCallback(
-    (e: React.SyntheticEvent<HTMLTextAreaElement>) => {
-      const newValue = +e.currentTarget.value;
-      if (isFinite(newValue)) {
-        onChange(newValue);
-      }
-    },
-    [onChange]
-  );
-
-  const onKeyDown = React.useCallback(
-    (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-      switch (e.keyCode) {
-        case KEY_UP:
-          e.preventDefault();
-          return onChange(value + 1);
-        case KEY_DOWN:
-          e.preventDefault();
-          return onChange(value - 1);
-        default:
-        // Nada
-      }
-    },
-    [value, onChange]
-  );
-
   return (
     <>
       <Label title={label} htmlFor={id}>
         {label}
       </Label>
       <ValueContainer>
-        <TextInputContainer focused={focused}>
-          <TextContainer>
-            <TextMirror style={{ opacity: focused ? 0 : 1 }}>
-              {value}
-            </TextMirror>
-            <TextField
-              rows={1}
-              id={id}
-              value={value}
-              onChange={onInputChange}
-              onFocus={onFocus}
-              onBlur={onBlur}
-              onKeyDown={onKeyDown}
-              style={{ opacity: focused ? 1 : 0 }}
-            />
-          </TextContainer>
-        </TextInputContainer>
+        <NumberInput id={id} value={value} onChange={onChange} />
       </ValueContainer>
     </>
   );
